requirements.txt
setuptools
streamlit
streamlit-audiorecorder
requests
python-dotenv
Flask
openai